Rewrite syncer TestLoadFork to use chain builder and fakes
