+++ chain/testing.go
@@ -175,9 +175,8 @@ func (f *Builder) Build(parent types.TipSet, width int, build func(b *BlockBuild
 		}
 
 		// Compute state root for this block.
-		tipSet, err := types.NewTipSet(b)
-		require.NoError(f.t, err)
-		b.StateRoot = f.ComputeState(tipSet)
+		prevState := f.StateForKey(parent.Key())
+		b.StateRoot, err = f.stateBuilder.ComputeState(prevState, [][]*types.SignedMessage{b.Messages})
 		require.NoError(f.t, err)
 
 		f.blocks[b.Cid()] = b
@@ -208,7 +207,7 @@ func (f *Builder) ComputeState(tip types.TipSet) cid.Cid {
 	require.NoError(f.t, err)
 	// Load the state of the parent tipset and compute the required state (recursively).
 	prev := f.StateForKey(parentKey)
-	state, err := f.stateBuilder.ComputeState(prev, tip)
+	state, err := f.stateBuilder.ComputeState(prev, tipMessages(tip))
 	require.NoError(f.t, err)
 	return state
 }
@@ -259,7 +258,7 @@ func (bb *BlockBuilder) SetStateRoot(root cid.Cid) {
 
 // StateBuilder abstracts the computation of state root CIDs from the chain builder.
 type StateBuilder interface {
-	ComputeState(prev cid.Cid, tip types.TipSet) (cid.Cid, error)
+	ComputeState(prev cid.Cid, blocksMessages [][]*types.SignedMessage) (cid.Cid, error)
 	Weigh(tip types.TipSet, state cid.Cid) (uint64, error)
 }
 
@@ -268,16 +267,16 @@ type FakeStateBuilder struct {
 }
 
 // ComputeState computes a fake state from a previous state root CID and the messages contained
-// in a tipset. Note that if the tipset is empty of messages, the resulting state is the same
-// as the input state.
+// in list-of-lists of messages in blocks. Note that if there are no messages, the resulting state
+// is the same as the input state.
 // This differs from the true state transition function in that messages that are duplicated
 // between blocks in the tipset are not ignored.
-func (FakeStateBuilder) ComputeState(prev cid.Cid, tip types.TipSet) (cid.Cid, error) {
+func (FakeStateBuilder) ComputeState(prev cid.Cid, blocksMessages [][]*types.SignedMessage) (cid.Cid, error) {
 	// Accumulate the cids of the previous state and of all messages in the tipset.
 	inputs := []cid.Cid{prev}
-	for i := 0; i < tip.Len(); i++ {
-		for j := 0; j < len(tip.At(i).Messages); j++ {
-			mCId, err := tip.At(i).Messages[j].Cid()
+	for _, blockMessages := range blocksMessages {
+		for _, msg := range blockMessages {
+			mCId, err := msg.Cid()
 			if err != nil {
 				return cid.Undef, err
 			}
@@ -313,8 +312,8 @@ type FakeStateEvaluator struct {
 }
 
 // RunStateTransition delegates to StateBuilder.ComputeState.
-func (e *FakeStateEvaluator) RunStateTransition(ctx context.Context, ts types.TipSet, ancestors []types.TipSet, stateID cid.Cid) (cid.Cid, error) {
-	return e.ComputeState(stateID, ts)
+func (e *FakeStateEvaluator) RunStateTransition(ctx context.Context, tip types.TipSet, ancestors []types.TipSet, stateID cid.Cid) (cid.Cid, error) {
+	return e.ComputeState(stateID, tipMessages(tip))
 }
 
 // IsHeavier compares chains weighed with StateBuilder.Weigh.
@@ -421,4 +420,12 @@ func makeCid(i interface{}) (cid.Cid, error) {
 		MhType:   types.DefaultHashFunction,
 		MhLength: -1,
 	}.Sum(bytes)
+}
+
+func tipMessages(tip types.TipSet) [][]*types.SignedMessage {
+	var msgs [][]*types.SignedMessage
+	for i := 0; i < tip.Len(); i++ {
+		msgs = append(msgs, tip.At(i).Messages)
+	}
+	return msgs
 }
